fix(v3-compiler): ProcessExit.ExitFailure.code uses i64 (substrate Int width)
Codex #1540: align host mirror with LiteralBits::Int / v3 Int carrier; drop i32 narrowing.

+++ src/v3/compiler/src/process_exit.rs
@@ -10,9 +10,10 @@
 
 // Practice 4 (coproduct checkpoint, `docs/modeling-discipline.md` §P1): 🟢 GREEN —
 // terminal host mirror of `dsl/std/process.dag` `ProcessExit`; no extra variants
-// or semantics beyond that substrate coproduct.
+// or semantics beyond that substrate coproduct. `ExitFailure.code` uses `i64`
+// — the same host width as `LiteralBits::Int` / substrate `Int` in `dag.rs`.
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub enum ProcessExit {
     ExitSuccess,
-    ExitFailure { code: i32, reason: String },
+    ExitFailure { code: i64, reason: String },
 }
